feat: count radler as 0.5 beer in leaderboard (#215)
* feat: count radler as 0.5 beer in leaderboard + backfill script

- Update get_global_leaderboard and get_group_leaderboard RPCs to
  weight radler as 0.5 and beer as 1.0 using SUM(CASE...)
- Change total_beers return type from BIGINT to NUMERIC
- Update leaderboard schema to accept decimal totalBeers
- Add parseFloat in repository mapper for NUMERIC strings
- Add scoring note subtitle to web and mobile leaderboard UI
- Add backfill script for Starkbierfest 2026 attendances with
  beer_count > 0 but no consumptions (run manually on production)
- Regenerate OpenAPI spec and API client types


* fix: address Copilot review - DROP before CREATE, search_path, ORDER BY

- DROP FUNCTION before CREATE (not OR REPLACE) since return type changes
  from BIGINT to NUMERIC — Postgres requires this
- Add SET search_path = public on SECURITY DEFINER function
- Add ORDER BY to backfill script's LIMIT 1 festival query


* test: add decimal totalBeers test for radler weighting

Verifies that the API correctly handles decimal total_beers values
(returned as strings by Supabase NUMERIC type) and serializes them
properly in the response (e.g., 3.5 for 3 beers + 1 radler).


* chore: remove backfill script (already run on production)


* refactor: CTE, consistent parseFloat, group scoring note

- Refactor get_group_leaderboard to use CTE (like global), eliminating
  4x repeated CASE WHEN weighting expression
- Fix inconsistent parseFloat pattern: use `parseFloat(x || "0")` to
  match existing avg_beers convention
- Add scoring note subtitle to group leaderboard pages (web + mobile)
  since they also use the weighted RPCs

+++ packages/shared/src/schemas/leaderboard.schema.ts
@@ -9,7 +9,7 @@ export const LeaderboardEntrySchema = z.object({
   fullName: z.string().nullable(),
   avatarUrl: z.url().nullable(),
   daysAttended: z.number().int(),
-  totalBeers: z.number().int(),
+  totalBeers: z.number(),
   avgBeers: z.number(),
   position: z.number().int(),
   groupCount: z.number().int().optional(),
